+++ src/AudioPlayer.cpp
@@ -736,6 +736,15 @@ void AudioPlayer_VolumeToQueueSender(const int32_t _newVolume, bool reAdjustRota
 // Receives de-serialized RFID-data (from NVS) and dispatches playlists for the given
 // playmode to the track-queue.
 void AudioPlayer_TrackQueueDispatcher(const char *_itemToPlay, const uint32_t _lastPlayPos, const uint32_t _playMode, const uint16_t _trackLastPlayed) {
+    // Make sure last playposition for audiobook is saved when new RFID-tag is applied
+    #ifdef SAVE_PLAYPOS_WHEN_RFID_CHANGE
+        if (!gPlayProperties.pausePlay && (gPlayProperties.playMode == AUDIOBOOK || gPlayProperties.playMode == AUDIOBOOK_LOOP)) {
+            AudioPlayer_TrackControlToQueueSender(PAUSEPLAY);
+            while (!gPlayProperties.pausePlay) {    // Make sure to wait until playback is paused in order to be sure that playposition saved in NVS
+                vTaskDelay(portTICK_RATE_MS * 100u);
+            }
+        }
+    #endif
     char *filename;
     filename = (char *) x_malloc(sizeof(char) * 255);
 
@@ -763,12 +772,26 @@ void AudioPlayer_TrackQueueDispatcher(const char *_itemToPlay, const uint32_t _l
     if (musicFiles == NULL) {
         Log_Println((char *) FPSTR(errorOccured), LOGLEVEL_ERROR);
         System_IndicateError();
-        gPlayProperties.playMode = NO_PLAYLIST;
+        if (!gPlayProperties.pausePlay) {
+            AudioPlayer_TrackControlToQueueSender(STOP);
+            while (!gPlayProperties.pausePlay) {
+                vTaskDelay(portTICK_RATE_MS * 10u);
+            }
+        } else {
+            gPlayProperties.playMode = NO_PLAYLIST;
+        }
         return;
     } else if (!strcmp(*(musicFiles - 1), "0")) {
         Log_Println((char *) FPSTR(noMp3FilesInDir), LOGLEVEL_NOTICE);
         System_IndicateError();
-        gPlayProperties.playMode = NO_PLAYLIST;
+        if (!gPlayProperties.pausePlay) {
+            AudioPlayer_TrackControlToQueueSender(STOP);
+            while (!gPlayProperties.pausePlay) {
+                vTaskDelay(portTICK_RATE_MS * 10u);
+            }
+        } else {
+            gPlayProperties.playMode = NO_PLAYLIST;
+        }
         free(filename);
         return;
     }